Restructure sampling for Zq
